requirements.txt
appdirs==1.4.2
decorator==4.0.11
Django==1.8.16
django-bitfield==1.8.0
django-common-helpers==0.9.0
django-cron==0.5.0
django-dirtyfields==1.2.1
django-grappelli==2.7.3
djangorestframework==3.5.3
filelock==2.0.7
httplib2==0.10.3
ipython==5.1.0
ipython-genutils==0.1.0
mysqlclient==1.3.9
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
pyTelegramBotAPI==2.3.0
pytz==2016.10
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
Trolly==0.2.2
wcwidth==0.1.7
